Move App Insights configuration to a separate file
