+++ FoxHound.Client/src/App.tsx
@@ -4,27 +4,9 @@ import React from "react";
 import ReactDOM from "react-dom";
 import NavDrawer from "./Navigation/NavDrawer";
 import { BrowserRouter as Router } from "react-router-dom";
-import { ApplicationInsights } from "@microsoft/applicationinsights-web";
-import {
-  ReactPlugin,
-  withAITracking,
-} from "@microsoft/applicationinsights-react-js";
-import { createBrowserHistory } from "history";
+import withAppInsights from "./Utility/AppInsights";
 
-const browserHistory = createBrowserHistory({ basename: "" });
-const reactPlugin = new ReactPlugin();
-const appInsights = new ApplicationInsights({
-  config: {
-    instrumentationKey: process.env.APP_INSIGHTS_INSTRUMENTATION_KEY,
-    extensions: [reactPlugin],
-    extensionConfig: {
-      [reactPlugin.identifier]: { history: browserHistory },
-    },
-  },
-});
-appInsights.loadAppInsights();
-
-const App: React.ComponentClass = withAITracking(reactPlugin, () => {
+const App: React.ComponentClass = withAppInsights(() => {
   return (
     <Router>
       <NavDrawer />
