+++ FoxHound.Client/src/Utility/AppInsights.ts
@@ -1,0 +1,23 @@
+import { ApplicationInsights } from "@microsoft/applicationinsights-web";
+import {
+  ReactPlugin,
+  withAITracking,
+} from "@microsoft/applicationinsights-react-js";
+import { createBrowserHistory } from "history";
+
+const browserHistory = createBrowserHistory({ basename: "" });
+const reactPlugin = new ReactPlugin();
+const ai = new ApplicationInsights({
+  config: {
+    instrumentationKey: process.env.APP_INSIGHTS_INSTRUMENTATION_KEY,
+    extensions: [reactPlugin],
+    extensionConfig: {
+      [reactPlugin.identifier]: { history: browserHistory },
+    },
+  },
+});
+ai.loadAppInsights();
+
+export default (Component: React.FC) => withAITracking(reactPlugin, Component);
+export const appInsights = ai.appInsights;
+export { reactPlugin };
